+++ spec/features/backup_spec.rb
@@ -1,39 +1,42 @@
 require "features/helper"
 
 RSpec.describe "backup", type: :aruba, docker: true do
-  include_context "account fixture"
   include_context "message-fixtures"
 
   let(:backup_folders) { [{name: folder}] }
   let(:folder) { "my-stuff" }
   let(:messages_as_mbox) do
-    message_as_mbox_entry(msg1) + message_as_mbox_entry(msg2)
+    to_mbox_entry(**msg1) + to_mbox_entry(**msg2)
   end
+  let(:account) { test_server_connection_parameters }
+  let(:email) { account[:username] }
+  let(:write_config) { create_config(accounts: [account]) }
 
   let!(:pre) do
-    server_delete_folder folder
+    test_server.delete_folder folder
   end
+  let(:command) { "imap-backup backup" }
   let!(:setup) do
-    server_create_folder folder
-    send_email folder, msg1
-    send_email folder, msg2
-    create_config(accounts: [account.to_h])
+    test_server.create_folder folder
+    test_server.send_email folder, **msg1
+    test_server.send_email folder, **msg2
+    write_config
 
-    run_command_and_stop("imap-backup backup")
+    run_command_and_stop command
   end
 
   after do
-    server_delete_folder folder
-    disconnect_imap
+    test_server.delete_folder folder
+    test_server.disconnect
   end
 
   it "downloads messages" do
-    expect(mbox_content(folder)).to eq(messages_as_mbox)
+    expect(mbox_content(email, folder)).to eq(messages_as_mbox)
   end
 
   describe "IMAP metadata" do
-    let(:imap_metadata) { imap_parsed(folder) }
-    let(:folder_uids) { server_uids(folder) }
+    let(:imap_metadata) { imap_parsed(email, folder) }
+    let(:folder_uids) { test_server.folder_uids(folder) }
 
     it "saves IMAP metadata in a JSON file" do
       expect { imap_metadata }.to_not raise_error
@@ -51,71 +54,120 @@ RSpec.describe "backup", type: :aruba, docker: true do
 
     it "records message offsets in the mbox file" do
       offsets = imap_metadata[:messages].map { |m| m[:offset] }
-      expected = [0, message_as_mbox_entry(msg1).length]
+      expected = [0, to_mbox_entry(**msg1).length]
 
       expect(offsets).to eq(expected)
     end
 
     it "records message lengths in the mbox file" do
       lengths = imap_metadata[:messages].map { |m| m[:length] }
-      expected = [message_as_mbox_entry(msg1).length, message_as_mbox_entry(msg2).length]
+      expected = [to_mbox_entry(**msg1).length, to_mbox_entry(**msg2).length]
 
       expect(lengths).to eq(expected)
     end
 
     it "records uid_validity" do
-      expect(imap_metadata[:uid_validity]).to eq(server_uid_validity(folder))
+      expect(imap_metadata[:uid_validity]).to eq(test_server.folder_uid_validity(folder))
+    end
+
+    context "with the --refresh option" do
+      let(:command) { "imap-backup backup --refresh" }
+      let(:connection) { Imap::Backup::Account::Connection.new(Imap::Backup::Account.new(account)) }
+
+      context "with messages that have already been backed up" do
+        let!(:pre) do
+          super()
+          write_config
+          test_server.create_folder folder
+          test_server.send_email folder, **msg3, flags: [:Draft]
+          connection.run_backup
+          connection.disconnect
+          test_server.set_flags folder, [1], [:Seen]
+        end
+
+        it "updates flags" do
+          imap_content = imap_parsed(email, folder)
+          message3 = imap_content[:messages].first
+          flags = message3[:flags].reject { |f| f == "Recent" }
+          expect(flags).to eq(["Seen"])
+        end
+      end
     end
 
     context "when uid_validity does not match" do
       let(:new_name) { "NEWNAME" }
-      let(:original_folder_uid_validity) { server_uid_validity(folder) }
-      let(:connection) { Imap::Backup::Account::Connection.new(account) }
+      let(:original_folder_uid_validity) { test_server.folder_uid_validity(folder) }
+      let(:connection) { Imap::Backup::Account::Connection.new(Imap::Backup::Account.new(account)) }
       let!(:pre) do
         super()
-        server_delete_folder new_name
-        server_create_folder folder
-        send_email folder, msg3
+        test_server.delete_folder new_name
+        test_server.create_folder folder
+        test_server.send_email folder, **msg3
         original_folder_uid_validity
         connection.run_backup
         connection.disconnect
-        server_rename_folder folder, new_name
+        test_server.rename_folder folder, new_name
       end
       let(:renamed_folder) { "#{folder}-#{original_folder_uid_validity}" }
 
       after do
-        server_delete_folder new_name
+        test_server.delete_folder new_name
       end
 
       it "renames the old backup" do
-        expect(mbox_content(renamed_folder)).to eq(message_as_mbox_entry(msg3))
+        expect(mbox_content(email, renamed_folder)).to eq(to_mbox_entry(**msg3))
       end
 
       it "renames the old metadata file" do
-        expect(imap_parsed(renamed_folder)).to be_a Hash
+        expect(imap_parsed(email, renamed_folder)).to be_a Hash
       end
 
       it "downloads messages" do
-        expect(mbox_content(folder)).to eq(messages_as_mbox)
+        expect(mbox_content(email, folder)).to eq(messages_as_mbox)
       end
 
       it "creates a metadata file" do
-        expect(imap_parsed(folder)).to be_a Hash
+        expect(imap_parsed(email, folder)).to be_a Hash
       end
 
       context "when a renamed local backup exists" do
         let!(:pre) do
           super()
-          create_directory local_backup_path
+          create_directory account[:local_path]
           valid_imap_data = {version: 3, uid_validity: 1, messages: []}
-          File.write(imap_path(renamed_folder), valid_imap_data.to_json)
-          File.write(mbox_path(renamed_folder), "existing mbox")
+          imap_path = File.join(account[:local_path], "#{renamed_folder}.imap")
+          File.write(imap_path, valid_imap_data.to_json)
+          mbox_path = File.join(account[:local_path], "#{renamed_folder}.mbox")
+          File.write(mbox_path, "existing mbox")
         end
 
         it "renames the renamed backup to a uniquely name" do
           renamed = "#{folder}-#{original_folder_uid_validity}-1"
-          expect(mbox_content(renamed)).to eq(message_as_mbox_entry(msg3))
+          expect(mbox_content(email, renamed)).to eq(to_mbox_entry(**msg3))
         end
+      end
+    end
+  end
+
+  context "in mirror mode" do
+    let(:account) { test_server_connection_parameters.merge(mirror_mode: true) }
+    let(:imap_path) { File.join(account[:local_path], "Foo.imap") }
+    let(:mbox_path) { File.join(account[:local_path], "Foo.mbox") }
+
+    let!(:pre) do
+      create_directory account[:local_path]
+      valid_imap_data = {version: 3, uid_validity: 1, messages: [{uid: 1, offset: 0, length: 3}]}
+      File.write(imap_path, valid_imap_data.to_json)
+      File.write(mbox_path, "existing mbox")
+    end
+
+    context "with folders that are not being backed up" do
+      it "deletes .imap files" do
+        expect(File.exist?(imap_path)).to be false
+      end
+
+      it "deletes .mbox files" do
+        expect(File.exist?(mbox_path)).to be false
       end
     end
   end