+++ spec/features/local/list_emails_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe "Listing emails", type: :aruba do
 
   before do
     create_config(accounts: [account])
-    store_email(email: email, folder: "my_folder", subject: "Ciao")
+    append_local(email: email, folder: "my_folder", subject: "Ciao")
     run_command_and_stop("imap-backup local list #{email} my_folder")
   end